Add support to resolve config values from filepath
Config files contains a mixture of sensitive and non-sensitive values.
Non-sensitive values can be hardcoded into the config.yaml file and can
be git committed or can be mounted into the application without
considering the risk factor.

In case of sensitive values, there are two ways to mount the respective
values: as environment or as volume mount. `viper` out of the box
supports marking the field as `env|KEY` and respective value can be
picked from the environment variables.

In this commit, we are adding support to fetch the secret value from the
file by setting the value as `file|/path/to/file` and during runtime,
resolving those values.

Also adding the cursor rule for future reference.

Signed-off-by: vinamra28 <vinjain@redhat.com>