Fix parsing of secrets and lint issues
If the default config contains some wildcards which are being overriden
in the environment specific config, while parsing the default config,
code was trying to resolve the default config first and then the
environment specific config, thus failing the config parsing.

Rearranging the config parsing and variable substitution in the order
of: parse default config, parse env specific config and then resolve the
variables which aren't already parsed.

Signed-off-by: vinamra28 <vinjain@redhat.com>